提交

+++ app/Services/Impl/SpecsValueServiceImpl.php
@@ -85,6 +85,9 @@ class SpecsValueServiceImpl implements SpecsService
         if (!$specsValue) {
             throw new \Exception('规格属性不存在');
         }
+        if (!$specsValue->delete_time != 0) {
+            throw new \Exception('规格属性已删除');
+        }
         unset($fields['id']);
         $this->model->updateById($id, $fields);
     }
@@ -94,6 +97,9 @@ class SpecsValueServiceImpl implements SpecsService
         $specsValue = $this->model->find($id);
         if (!$specsValue) {
             throw new \Exception('规格属性不存在');
+        }
+        if (!$specsValue->delete_time != 0) {
+            throw new \Exception('规格属性已删除');
         }
         $this->model->deleteById($id);
     }
